fix: using uint64 pointer to fix unalign problem for x86 platform

+++ buffer_test.go
@@ -49,7 +49,7 @@ func TestX(t *testing.T) {
 }
 
 func TestBuffer(t *testing.T) {
-	buf := newRingBuffer[uint64](1024)
+	buf := newRingBuffer[uint8](1024)
 	buf.write(0, 1)
 	x := buf.element(0)
 	fmt.Println(x)
